refactor: separate zabo component
