+++ src/components/zabo/Zabo.tsx
@@ -2,12 +2,11 @@ import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 import { Link } from "react-router-dom";
 
-import DefaultProfile from "@/assets/default-profile.svg?react";
 import { Notice } from "@/types/interfaces";
-
-import DDay from "../DDay/DDay";
-import ZaboActions from "./ZaboActions";
-import ZaboTags from "./ZaboTags";
+import ZaboActions from "./zaboActions/ZaboActions";
+import ZaboTags from "./zaboTags/ZaboTags";
+import ZaboHeader from "./zaboHeader/ZaboHeader";
+import ZaboImages from "./zaboImages/ZaboImages";
 
 export type ZaboOrigin = "width" | "height";
 
@@ -28,8 +27,6 @@ const Zabo = (props: ZaboProps) => {
   const { createdAt, author, deadline, reactions, title, imageUrls, tags, id } =
     props;
 
-  const timeAgo = dayjs(createdAt).fromNow();
-
   const hasImage = imageUrls.length > 0;
 
   return (
@@ -39,26 +36,7 @@ const Zabo = (props: ZaboProps) => {
           "flex flex-col rounded-[10px] py-[10px] text-text transition hover:bg-greyLight dark:hover:bg-d_greyDark"
         }
       >
-        <div className={"mx-3 my-[10px] flex flex-wrap items-center gap-y-3"}>
-          <DefaultProfile width={36} height={36} />
-
-          <p className={"ml-2 text-lg font-medium dark:text-d_white"}>
-            {author.name}
-          </p>
-
-          <p className={"mx-[5px] font-bold text-greyDark dark:text-grey"}>·</p>
-
-          <p className={"font-medium text-greyDark dark:text-grey"}>
-            {timeAgo}
-          </p>
-
-          {deadline !== null && (
-            <>
-              <div className="w-[15px]" />
-              <DDay deadline={dayjs(deadline)} />
-            </>
-          )}
-        </div>
+        <ZaboHeader author={author} deadline={deadline} createdAt={createdAt} />
 
         <p className={"mx-4 mb-[10px] text-xl font-semibold dark:text-d_white"}>
           {title}
@@ -66,26 +44,7 @@ const Zabo = (props: ZaboProps) => {
 
         {!hasImage && <ZaboTags notice={props} />}
 
-        {hasImage && (
-          <div className={"flex justify-center"}>
-            {/* // 아래 div에 바로 justify-center를 적용하면, 1번 2번 이미지가 잘려서
-          보임 */}
-            <div className={"flex flex-nowrap gap-[10px] overflow-x-scroll"}>
-              {imageUrls.map((url) => (
-                <img
-                  key={url}
-                  src={url}
-                  alt={title}
-                  width={200}
-                  height={200}
-                  className={
-                    "rounded-[5px] border border-gray-300 object-cover"
-                  }
-                />
-              ))}
-            </div>
-          </div>
-        )}
+        {hasImage && <ZaboImages title={title} imageUrls={imageUrls} />}
 
         {hasImage && (
           <div className="mx-2 my-4">
